split feature file. added endpoint placeholder, added tests, recased feature files
