+++ features/eps_fhir/dispenseNotification.feature
@@ -1,0 +1,17 @@
+@eps_fhir @smoke @regression @blocker @eps_prescribing_dispensing @create
+
+Feature: I can manage dispense notifications
+@allure.tms:https://nhsd-jira.digital.nhs.uk/browse/AEA-3868
+Scenario: I can amend a single dispense notification
+Given a new prescription has been dispensed
+When I amend the dispense notification
+Then the response indicates a success
+And the response body indicates a successful amend dispense action
+
+@withdraw
+@allure.tms:https://nhsd-jira.digital.nhs.uk/browse/AEA-3867
+Scenario: I can withdraw a dispense notification
+Given a new prescription has been dispensed
+When I withdraw the dispense notification
+Then the response indicates a success
+And the response body indicates a successful dispense withdrawal action
